Add index buffer binding to PxeVertexArray

+++ PXEngine/include/PxeVertexBufferFormat.h
@@ -7,7 +7,6 @@
 
 namespace pxengine {
 	// Helper class to manage PxeVertexBuffer and PxeVertexArray data format/layout
-	// Warning this class is NOT thread safe
 	class PxeVertexBufferFormat
 	{
 	public:
@@ -39,15 +38,15 @@ namespace pxengine {
 		// Inserts an PxeVertexBufferAttrib and returns the index it was inserted at
 		// Note: attrib Offset value is ignored and recalculated when inserted and does not change the passed attrib Offset value
 		// Note: the return is not a size_t because glVertexAttribPointer only supports a GLuint
-		uint32_t addAttrib(const PxeVertexBufferAttrib& attrib) {
+		uint8_t addAttrib(const PxeVertexBufferAttrib& attrib) {
 			PxeVertexBufferAttrib& a = _attribs.emplace_back(attrib);
 			a.Offset = getStride();
 			recalculateStride();
-			return static_cast<uint32_t>(_attribs.size() - 1);
+			return static_cast<uint8_t>(_attribs.size() - 1);
 		}
 
 		// Returns true if {index} was valid and {attrib} was set to the current value at {index}
-		PXE_NODISCARD bool getAttrib(size_t index, PxeVertexBufferAttrib& attrib) const {
+		PXE_NODISCARD bool getAttrib(uint8_t index, PxeVertexBufferAttrib& attrib) const {
 			if (index >= _attribs.size()) return false;
 			attrib = _attribs[index];
 			return true;
@@ -57,8 +56,7 @@ namespace pxengine {
 		inline PXE_NODISCARD uint32_t getStride() const { return _stride; }
 
 		// Returns the number of attribs stored 
-		// Note: the return is not a size_t because glVertexAttribPointer only supports a GLuint
-		PXE_NODISCARD uint32_t getAttribCount() const { return static_cast<uint32_t>(_attribs.size()); }
+		PXE_NODISCARD uint8_t getAttribCount() const { return static_cast<uint8_t>(_attribs.size()); }
 
 	private:
 		// Recalculates the _stride value
